+++ app/models/epoch_statistic.rb
@@ -1,5 +1,5 @@
 class EpochStatistic < ApplicationRecord
-  VALID_INDICATORS = %w(difficulty uncle_rate hash_rate).freeze
+  VALID_INDICATORS = %w(difficulty uncle_rate hash_rate epoch_time epoch_length).freeze
 end
 
 # == Schema Information