+++ app/serializers/ckb_transactions_serializer.rb
@@ -14,14 +14,40 @@ class CkbTransactionsSerializer
   end
 
   attribute :display_inputs do |object, params|
-    params && params[:previews] ? object.display_inputs(previews: true) : object.display_inputs
+    if params && params[:previews]
+      if object.display_inputs_info.present?
+        object.display_inputs_info(previews: true)
+      else
+        object.display_inputs(previews: true)
+      end
+    else
+      object.display_inputs_info.presence || object.display_inputs
+    end
   end
 
   attribute :display_outputs do |object, params|
-    params && params[:previews] ? object.display_outputs(previews: true) : object.display_outputs
+    if params && params[:previews]
+      if object.display_inputs_info.present?
+        object.display_outputs_info(previews: true)
+      else
+        object.display_outputs(previews: true)
+      end
+    else
+      if object.display_inputs_info.present?
+        object.display_outputs_info
+      else
+        object.display_outputs
+      end
+    end
   end
 
   attribute :income do |object, params|
-    params && params[:previews] && params[:address] ? object.income(params[:address]) : nil
+    if params && params[:previews] && params[:address].present?
+      if object.tx_display_info.present?
+        object.tx_display_info.income[params[:address].address_hash]
+      else
+        object.income(params[:address])
+      end
+    end
   end
 end