+++ db/migrate/20201218065319_add_index_on_created_at_unixtimestamp_to_daily_statistics.rb
@@ -1,0 +1,5 @@
+class AddIndexOnCreatedAtUnixtimestampToDailyStatistics < ActiveRecord::Migration[6.0]
+  def change
+    add_index :daily_statistics, :created_at_unixtimestamp, order: { created_at_unixtimestamp: "DESC NULLS LAST" }
+  end
+end