+++ db/migrate/20210304115516_add_index_on_scripts.rb
@@ -1,0 +1,6 @@
+class AddIndexOnScripts < ActiveRecord::Migration[6.0]
+  def change
+    add_index :lock_scripts, :script_hash
+    add_index :type_scripts, :script_hash
+  end
+end