+++ test/controllers/api/v1/address_dao_transactions_controller_test.rb
@@ -208,25 +208,6 @@ module Api
         assert_equal response_transaction, response.body
       end
 
-      test "should return pagination links in response body" do
-        page = 2
-        page_size = 3
-        address = create(:address, :with_transactions, transactions_count: 30)
-        fake_dao_deposit_transaction(30, address)
-
-        links = {
-          self: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page=2&page_size=3",
-          first: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page_size=3",
-          prev: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page_size=3",
-          next: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page=3&page_size=3",
-          last: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page=10&page_size=3"
-        }
-
-        valid_get api_v1_address_dao_transaction_url(address.address_hash), params: { page: page, page_size: page_size }
-
-        assert_equal links.stringify_keys.sort, json["links"].sort
-      end
-
       test "should return meta that contained total in response body" do
         address = create(:address, :with_transactions, transactions_count: 3)
         fake_dao_deposit_transaction(3, address)
@@ -234,17 +215,6 @@ module Api
         valid_get api_v1_address_dao_transaction_url(address.address_hash)
 
         assert_equal 3, json.dig("meta", "total")
-      end
-
-      test "should return pagination links that only contain self in response bod when there is no blocks" do
-        address = create(:address, :with_transactions)
-
-        links = {
-          self: "#{api_v1_address_dao_transaction_url(address.address_hash)}?page_size=10"
-        }
-
-        valid_get api_v1_address_dao_transaction_url(address.address_hash)
-        assert_equal links.stringify_keys.sort, json["links"].sort
       end
     end
   end