+++ test/controllers/api/v1/address_udt_transactions_controller_test.rb
@@ -200,7 +200,7 @@ module Api
         page_size = 5
         udt = create(:udt, published: true)
         address = create(:address, :with_udt_transactions, transactions_count: 30, udt: udt)
-        address_udt_transactions = address.ckb_udt_transactions(udt.id).order(block_timestamp: :desc).page(page).per(page_size)
+        address_udt_transactions = address.ckb_udt_transactions(udt.id).order("block_timestamp desc nulls last, id desc").page(page).per(page_size)
 
         valid_get api_v1_address_udt_transaction_url(address.address_hash, type_hash: udt.type_hash), params: { page: page, page_size: page_size }
         options = FastJsonapi::PaginationMetaGenerator.new(request: request, records: address_udt_transactions, page: page, page_size: page_size).call
@@ -233,18 +233,6 @@ module Api
         valid_get api_v1_address_udt_transaction_url(address.address_hash, type_hash: udt.type_hash)
 
         assert_equal 6, json.dig("meta", "total")
-      end
-
-      test "should return pagination links that only contain self in response body when there is no blocks" do
-        udt = create(:udt, published: true)
-        address = create(:address, udt: udt)
-
-        links = {
-          self: "#{api_v1_address_udt_transaction_url(address.address_hash, type_hash: udt.type_hash)}?page_size=10"
-        }
-
-        valid_get api_v1_address_udt_transaction_url(address.address_hash, type_hash: udt.type_hash)
-        assert_equal links.stringify_keys.sort, json["links"].sort
       end
     end
   end