+++ test/controllers/api/v1/contract_transactions_controller_test.rb
@@ -173,25 +173,6 @@ module Api
         assert_equal response_transaction, response.body
       end
 
-      test "should return pagination links in response body" do
-        address = create(:address)
-        fake_dao_deposit_transaction(30, address)
-        page = 2
-        page_size = 3
-
-        links = {
-          self: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page=2&page_size=3",
-          first: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page_size=3",
-          prev: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page_size=3",
-          next: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page=3&page_size=3",
-          last: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page=10&page_size=3"
-        }
-
-        valid_get api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME), params: { page: page, page_size: page_size }
-
-        assert_equal links.stringify_keys.sort, json["links"].sort
-      end
-
       test "should return meta that contained total in response body" do
         address = create(:address)
         fake_dao_deposit_transaction(3, address)
@@ -199,15 +180,6 @@ module Api
         valid_get api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)
 
         assert_equal 3, json.dig("meta", "total")
-      end
-
-      test "should return pagination links that only contain self in response bod when there is no transactions" do
-        links = {
-          self: "#{api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)}?page_size=10"
-        }
-
-        valid_get api_v1_contract_transaction_url(DaoContract::CONTRACT_NAME)
-        assert_equal links.stringify_keys.sort, json["links"].sort
       end
     end
   end